+++ main/PygameXtras/__init__.py
@@ -1,1 +1,1 @@
-from .src.classes import *
+from .src import *